+++ src/main/java/com/example/namo2/domain/moim/MoimController.java
@@ -3,8 +3,8 @@ package com.example.namo2.domain.moim;
 import com.example.namo2.domain.memo.MoimMemoService;
 import com.example.namo2.domain.moim.dto.*;
 import com.example.namo2.global.common.response.BaseResponse;
-import com.example.namo2.domain.schedule.dto.DiaryDto;
-import com.example.namo2.domain.schedule.dto.SliceDiaryDto;
+import com.example.namo2.domain.schedule.ui.dto.DiaryDto;
+import com.example.namo2.domain.schedule.ui.dto.SliceDiaryDto;
 import com.example.namo2.global.utils.Converter;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
