+++ src/main/java/com/example/namo2/domain/moim/MoimScheduleRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.example.namo2.domain.moim;
 
-import com.example.namo2.domain.schedule.dto.DiaryDto;
-import com.example.namo2.domain.schedule.dto.SliceDiaryDto;
+import com.example.namo2.domain.schedule.ui.dto.DiaryDto;
+import com.example.namo2.domain.schedule.ui.dto.SliceDiaryDto;
 import org.springframework.data.domain.Pageable;
 
 import java.time.LocalDateTime;
