--- src/main/java/com/example/namo2/domain/schedule/AlarmRepositoryCustom.java
@@ -1,4 +1,0 @@
-package com.example.namo2.domain.schedule;
-
-public interface AlarmRepositoryCustom {
-}
