+++ src/main/java/com/example/namo2/domain/schedule/application/ScheduleFacade.java
@@ -1,0 +1,4 @@
+package com.example.namo2.domain.schedule.application;
+
+public class ScheduleFacade {
+}
