+++ src/main/java/com/example/namo2/domain/schedule/application/impl/ScheduleService.java
@@ -1,6 +1,9 @@
-package com.example.namo2.domain.schedule;
+package com.example.namo2.domain.schedule.application.impl;
 
 import com.example.namo2.domain.category.dao.repository.CategoryRepository;
+import com.example.namo2.domain.schedule.dao.repository.AlarmRepository;
+import com.example.namo2.domain.schedule.dao.repository.ImageRepository;
+import com.example.namo2.domain.schedule.dao.repository.ScheduleRepository;
 import com.example.namo2.global.common.exception.BaseException;
 import com.example.namo2.domain.moim.domain.MoimSchedule;
 import com.example.namo2.domain.moim.domain.MoimScheduleAndUser;
@@ -9,17 +12,17 @@ import com.example.namo2.domain.schedule.domain.Period;
 import com.example.namo2.domain.moim.MoimScheduleAndUserRepository;
 import com.example.namo2.domain.moim.MoimScheduleRepository;
 import com.example.namo2.domain.moim.MoimService;
-import com.example.namo2.domain.schedule.dto.DiaryDto;
+import com.example.namo2.domain.schedule.ui.dto.DiaryDto;
 import com.example.namo2.domain.category.domain.Category;
 import com.example.namo2.domain.schedule.domain.Image;
 import com.example.namo2.domain.schedule.domain.Schedule;
 import com.example.namo2.domain.user.domain.User;
-import com.example.namo2.domain.schedule.dto.GetDiaryRes;
-import com.example.namo2.domain.schedule.dto.GetScheduleRes;
-import com.example.namo2.domain.schedule.dto.OnlyDiaryDto;
-import com.example.namo2.domain.schedule.dto.PostScheduleReq;
-import com.example.namo2.domain.schedule.dto.ScheduleIdRes;
-import com.example.namo2.domain.schedule.dto.SliceDiaryDto;
+import com.example.namo2.domain.schedule.ui.dto.GetDiaryRes;
+import com.example.namo2.domain.schedule.ui.dto.GetScheduleRes;
+import com.example.namo2.domain.schedule.ui.dto.OnlyDiaryDto;
+import com.example.namo2.domain.schedule.ui.dto.PostScheduleReq;
+import com.example.namo2.domain.schedule.ui.dto.ScheduleIdRes;
+import com.example.namo2.domain.schedule.ui.dto.SliceDiaryDto;
 import com.example.namo2.domain.user.UserRepository;
 import com.example.namo2.global.utils.Converter;
 import com.example.namo2.global.utils.FileUtils;
