+++ src/main/java/com/example/namo2/domain/schedule/dao/repository/AlarmRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.example.namo2.domain.schedule.dao.repository;
+
+public interface AlarmRepositoryCustom {
+}
