+++ src/main/java/com/example/namo2/domain/schedule/dao/repository/ImageRepository.java
@@ -1,4 +1,4 @@
-package com.example.namo2.domain.schedule;
+package com.example.namo2.domain.schedule.dao.repository;
 
 import com.example.namo2.domain.schedule.domain.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
