+++ src/main/java/com/example/namo2/domain/schedule/dao/repository/ScheduleRepository.java
@@ -1,4 +1,4 @@
-package com.example.namo2.domain.schedule;
+package com.example.namo2.domain.schedule.dao.repository;
 
 import com.example.namo2.domain.schedule.domain.Schedule;
 import org.springframework.data.jpa.repository.JpaRepository;
