+++ src/main/java/com/example/namo2/domain/schedule/dao/repository/ScheduleRepositoryCustom.java
@@ -1,12 +1,12 @@
-package com.example.namo2.domain.schedule;
+package com.example.namo2.domain.schedule.dao.repository;
 
 import com.example.namo2.domain.schedule.domain.Schedule;
 import com.example.namo2.domain.user.domain.User;
 import com.example.namo2.domain.moim.dto.MoimScheduleRes;
-import com.example.namo2.domain.schedule.dto.DiaryDto;
-import com.example.namo2.domain.schedule.dto.GetScheduleRes;
-import com.example.namo2.domain.schedule.dto.OnlyDiaryDto;
-import com.example.namo2.domain.schedule.dto.SliceDiaryDto;
+import com.example.namo2.domain.schedule.ui.dto.DiaryDto;
+import com.example.namo2.domain.schedule.ui.dto.GetScheduleRes;
+import com.example.namo2.domain.schedule.ui.dto.OnlyDiaryDto;
+import com.example.namo2.domain.schedule.ui.dto.SliceDiaryDto;
 import org.springframework.data.domain.Pageable;
 
 import java.time.LocalDateTime;
