+++ src/main/java/com/example/namo2/domain/schedule/ui/dto/PatchScheduleReq.java
@@ -1,15 +1,15 @@
-package com.example.namo2.domain.schedule.dto;
+package com.example.namo2.domain.schedule.ui.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import java.util.Set;
+import java.util.List;
 
 @Getter
 @AllArgsConstructor
-public class PostScheduleReq {
+public class PatchScheduleReq {
     @NotBlank
     private String name;
     @NotNull
@@ -18,7 +18,8 @@ public class PostScheduleReq {
     private Long endDate;
     @NotNull
     private Integer interval;
-    private Set<Integer> alarmDate;
+    private List<Integer> deleteAlarmDate;
+    private List<Integer> addAlarmDate;
     private Double x;
     private Double y;
     private String locationName;
