+++ src/main/java/com/example/namo2/domain/schedule/ui/dto/SliceDiaryDto.java
@@ -1,4 +1,4 @@
-package com.example.namo2.domain.schedule.dto;
+package com.example.namo2.domain.schedule.ui.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
